requirements.txt
# installing required libraries in my_env
transformers==4.35.2
torch==2.1.1
gradio==4.44.0
langchain==0.0.343
huggingface-hub==0.19.4
accelerate==0.26.1
python-dotenv==1.0.0
